llama-index/pyproject.toml
[project]
name = "llama-index-examples"
version = "1.0.0"
description = "LlamaIndex framework examples"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "openai>=1.60.0",
    "llama-index-tools-google>=0.6.2",
    "llama-index-workflows>=2.11.5",
    "llama-index-utils-workflow>=0.9.2",
    "llama-agents-server>=0.1.3",
]
